Reorganize some methods
